Refactor dependency injection to avoid import loop
It was quite common that settings.Settings would use one or more structs
from resources or services as part of its injector. This would lead to a
dependency loop when using the settings.Settings was used in a package
and settings.Settings had dependencies to the same package.

Moving the injectors each to separate packages reduces the chance of a
dependency loop.