Workaround, yaml lib is clobbering ignored fields
This change introduces a work around bug that seems to be in the yaml
library. The library seems to be intermittently clobbering fields that
have the ignore tag set `yaml:"-"`. This bug is hard to reproduce and
hence a workaround has been implemented to reset the clobbered fields.
When time permits further investigation will ensue and an upstream bug
filed once evidence can be coded in a concise format.

The changes are...

 * config.go: The workaround.

 * install.go: Add file stat checks to ensure the existence of the yaml
   files.

 * installcmd_test.go: Test install sub command for disk templates.

 * test/fixtures: Add new test fixtures to test installed disk
   templates.